requirements.txt
azure.storage==0.32
